pyproject.toml
[project]
name = "self-healing-agent"
version = "0.1.0"
description = "AI agent for self-healing Java Spring Boot apps and software bug assistance using ADK"
authors = [
    { name = "Your Name", email = "you@example.com" } 
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "google-adk>=1.3.0",
    "langchain",
    "langchain-pinecone",
    "langchain-google-vertexai",
    "langchain-community>=0.3.25",
    "pinecone>=3.0.0",
    "google-cloud-aiplatform",
    "google-cloud-pubsub>=2.17.0",
    "python-dotenv>=1.1.0",
    "pydantic",
    "stackapi>=0.3.1",
    "toolbox-core==0.1.0"
]

[project.scripts]
software_bug_assistant = "self_healing_agent.main_agent:main"

[tool.adk]
root-agent = "self_healing_agent.main_agent:root_agent"
description = "Root agent responsible for orchestrating the self-healing pipeline."

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = [ "." ]
include = [ "self_healing_agent*" ]
